e-commerce/backend/requirements.txt
Django==5.0.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.4.0
djangorestframework-simplejwt==5.3.1
django-rest-auth==0.9.5
djongo==1.3.6
pymongo==4.6.1
stripe==7.10.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2
python-dotenv==1.0.0
cryptography==42.0.5
sqlparse==0.4.4
